shoppingList-tilføjVare - ready for Pull Request

+++ composeApp/src/main/kotlin/com/example/a62550_foodapp/db/entity/RecipeItem.kt
@@ -31,5 +31,5 @@ data class RecipeItem(
     @ColumnInfo(name = "item_group_id")
     val itemGroupId: Int,
 
-    val quantity: Float
+    val quantity: Int
 )
